+++ src/main/java/org/thoughtslive/jenkins/plugins/jira/service/JiraService.java
@@ -18,7 +18,6 @@ import okhttp3.ResponseBody;
 import org.thoughtslive.jenkins.plugins.jira.Site;
 import org.thoughtslive.jenkins.plugins.jira.api.ResponseData;
 import org.thoughtslive.jenkins.plugins.jira.login.SigningInterceptor;
-import org.thoughtslive.jenkins.plugins.jira.model.CommentParameter;
 import retrofit2.Retrofit;
 import retrofit2.adapter.rxjava.RxJavaCallAdapterFactory;
 import retrofit2.converter.jackson.JacksonConverterFactory;
@@ -179,44 +178,20 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Object> addComment(final String issueIdorKey, final String comment) {
+  public ResponseData<Object> addComment(final String issueIdorKey, final Object input) {
     try {
       return parseResponse(jiraEndPoints
-          .addComment(issueIdorKey, ImmutableMap.builder().put("body", comment).build()).execute());
-    } catch (Exception e) {
-      return buildErrorResponse(e);
-    }
-  }
-
-  public ResponseData<Object> addComment(final String issueIdorKey, final String comment, final String roleVisibility) {
-    try {
-      return parseResponse(jiraEndPoints
-          .addComment(issueIdorKey, new CommentParameter(comment, roleVisibility)).execute());
+          .addComment(issueIdorKey, input).execute());
     } catch (Exception e) {
       return buildErrorResponse(e);
     }
   }
 
   public ResponseData<Object> updateComment(final String issueIdorKey, final String commentId,
-      final String comment) {
+      final Object input) {
     try {
       return parseResponse(
-          jiraEndPoints
-              .updateComment(issueIdorKey, commentId,
-                  ImmutableMap.builder().put("body", comment).build())
-              .execute());
-    } catch (Exception e) {
-      return buildErrorResponse(e);
-    }
-  }
-
-  public ResponseData<Object> updateComment(final String issueIdorKey, final String commentId,
-  final String comment, final String roleVisibility) {
-    try {
-      return parseResponse(
-          jiraEndPoints
-              .updateComment(issueIdorKey, commentId, new CommentParameter(comment, roleVisibility))
-              .execute());
+          jiraEndPoints.updateComment(issueIdorKey, commentId, input).execute());
     } catch (Exception e) {
       return buildErrorResponse(e);
     }