+++ src/main/java/org/thoughtslive/jenkins/plugins/jira/steps/EditCommentStep.java
@@ -6,9 +6,11 @@ import hudson.Extension;
 import hudson.Util;
 import java.io.IOException;
 import lombok.Getter;
+import com.google.common.collect.ImmutableMap;
 import org.jenkinsci.plugins.workflow.steps.StepContext;
 import org.jenkinsci.plugins.workflow.steps.StepExecution;
 import org.kohsuke.stapler.DataBoundConstructor;
+import org.kohsuke.stapler.DataBoundSetter;
 import org.thoughtslive.jenkins.plugins.jira.api.ResponseData;
 import org.thoughtslive.jenkins.plugins.jira.util.JiraStepDescriptorImpl;
 import org.thoughtslive.jenkins.plugins.jira.util.JiraStepExecution;
@@ -28,18 +30,24 @@ public class EditCommentStep extends BasicJiraStep {
   @Getter
   private final String commentId;
 
+  @Deprecated
   @Getter
   private final String comment;
 
   @Getter
-  private final String roleVisibility;
+  private Object input;
 
+  @Deprecated
   @DataBoundConstructor
-  public EditCommentStep(final String idOrKey, final String commentId, final String comment, final String roleVisibility) {
+  public EditCommentStep(final String idOrKey, final String commentId, final String comment) {
     this.idOrKey = idOrKey;
     this.commentId = commentId;
     this.comment = comment;
-    this.roleVisibility = roleVisibility;
+  }
+
+  @DataBoundSetter
+  public void setInput(Object input) {
+    this.input = input;
   }
 
   @Override
@@ -79,10 +87,16 @@ public class EditCommentStep extends BasicJiraStep {
       ResponseData<Object> response = verifyInput();
 
       if (response == null) {
-        logger.println("JIRA: Site - " + siteName + " - Updating comment: " + step.getComment()
-            + " on issue: " + step.getIdOrKey());
-        response =
-            jiraService.updateComment(step.getIdOrKey(), step.getCommentId(), step.getComment(), step.getRoleVisibility());
+        if (step.getComment() != null) {
+          logger.println("JIRA: Site - " + siteName + " - Updating comment (deprecated): " + step.getComment()
+              + " on issue: " + step.getIdOrKey());
+          response = jiraService.updateComment(step.getIdOrKey(), step.getCommentId(), 
+            ImmutableMap.builder().put("body", step.getComment()).build());
+        } else {
+          logger.println("JIRA: Site - " + siteName + " - Updating comment: " + step.getInput()
+              + " on issue: " + step.getIdOrKey());
+          response = jiraService.updateComment(step.getIdOrKey(), step.getCommentId(), step.getInput());
+        }
       }
 
       return logResponse(response);
@@ -95,8 +109,15 @@ public class EditCommentStep extends BasicJiraStep {
 
       if (response == null) {
         final String idOrKey = Util.fixEmpty(step.getIdOrKey());
-        final String comment = Util.fixEmpty(step.getComment());
         final String commentId = Util.fixEmpty(step.getCommentId());
+
+        if (step.getComment() != null && step.getInput() != null) {
+          errorMessage = "Use comment or either input.";
+        }
+
+        if (step.getComment() == null && step.getInput() == null) {
+          errorMessage = "You need to set at least comment or input.";
+        }
 
         if (idOrKey == null) {
           errorMessage = "idOrKey is empty or null.";
@@ -106,8 +127,8 @@ public class EditCommentStep extends BasicJiraStep {
           errorMessage = "commentId is empty or null.";
         }
 
-        if (comment == null) {
-          errorMessage = "comment is empty or null.";
+        if (step.getComment() != null && step.getComment().isEmpty()) {
+          errorMessage = "comment is empty.";
         }
 
         if (errorMessage != null) {