+++ hugo/content/changelog.md
@@ -2,11 +2,11 @@
 title = "Change Log"
 description = "Change Log."
 date = "2017-11-12"
-lastmodifierdisplayname = "Naresh Rayapati"
+lastmodifierdisplayname = "Benedikt Hr"
 +++
 
 * #### **1.4.5** (Unreleased)
-  * Allow to set a role visibility when creating or editing comments.
+  * [JENKINS-49394](https://issues.jenkins-ci.org/browse/JENKINS-49394) Make add/edit comment step more generic to set a visibility.
 
 * #### **1.4.4**
   * [JENKINS-53044](https://issues.jenkins-ci.org/browse/JENKINS-53044) Add step to retrieve jira server info.